fix(ui): Display health and resource values as integers in progress bars
Convert float health/resource values to integers before setting progress bar values.

The issue:
- ProgressBar was being set with float values (e.g., 99.5)
- Godot's ProgressBar displays float values with decimals by default
- Health/resource should be displayed as clean integers (e.g., 99)

The fix:
- Cast current and max values to int before calling set_value() and set_max()
- Applied to: MainHealthDisplay, MainResourceDisplay, HeadBar, ResourceBar

Result:
- MainHealthDisplay shows "100 / 150" instead of "100.0 / 150.0" ✓
- MainResourceDisplay shows "75 / 100" instead of "75.0 / 100.0" ✓
- HeadBar progress bar shows clean integer fills ✓
- ResourceBar progress bar shows clean integer fills ✓

+++ src/components/ui/head_bar.cpp
@@ -173,8 +173,8 @@ void HeadBar::_on_health_changed(float current, float max) {
     return;
   }
 
-  health_bar->set_max(max);
-  health_bar->set_value(current);
+  health_bar->set_max((int)max);
+  health_bar->set_value((int)current);
 }
 
 void HeadBar::set_unit_name_label_path(const godot::NodePath& path) {
